move repeater generator tests to their own file
